Make the camera move the `TransformComponent` as well

+++ src/ECS/ColliderComponent.cpp
@@ -4,7 +4,7 @@ void ColliderComponent::update() {
   if (this->entity->hasComponent<TransformComponent>()) {
     auto transform = this->entity->getComponentPtr<TransformComponent>();
 
-    collider.x = transform->position.x - Game::camera.x;
+    collider.x = static_cast<int>(transform->position.x - Game::camera.x);
     collider.y = transform->position.y;
     collider.w = transform->width * transform->scale;
     collider.h = transform->height * transform->scale;
